+++ src/pages/code.tsx
@@ -1,17 +1,7 @@
 import React, { useRef, useState } from 'react'
 import SEO from 'components/seo'
 import Editor from 'components/Editor'
-import {
-    IconArrowRight,
-    IconArrowUpRight,
-    IconCheck,
-    IconExternal,
-    IconFlask,
-    IconToggle,
-    IconTrends,
-    IconWarning,
-} from '@posthog/icons'
-import Input from 'components/OSForm/input'
+import { IconArrowUpRight, IconCheck, IconFlask, IconToggle, IconTrends, IconWarning } from '@posthog/icons'
 import OSButton from 'components/OSButton'
 import { Accordion } from 'components/RadixUI/Accordion'
 import { LOGOS, type LogoKey } from 'constants/logos'
@@ -23,94 +13,12 @@ import { IconPop } from 'components/Code/IconPop'
 import { SignalsCallout } from 'components/Code/SignalsCallout'
 import { FlowDiagram } from 'components/Code/FlowDiagram'
 import { DottedConnection } from 'components/Code/DottedConnection'
-import {
-    StickerAi,
-    StickerBulb,
-    StickerCloud,
-    StickerCoffee,
-    StickerELearning,
-    StickerMicroscope,
-    StickerTombstone,
-    StickerPause,
-    StickerPullRequest,
-    StickerRobot,
-    StickerTerminal,
-    StickerUsers,
-    StickerWebsite,
-    StickerZZZ,
-    StickerOne,
-} from 'components/Stickers/Stickers'
+import { StickerTombstone } from 'components/Stickers/Stickers'
 import { ZoomImage } from 'components/ZoomImage'
-import { DebugContainerQuery } from 'components/DebugContainerQuery'
-import usePostHog from '../hooks/usePostHog'
-import useProduct from '../hooks/useProduct'
-import { useApp } from '../context/App'
 import CloudinaryImage from 'components/CloudinaryImage'
 import Link from 'components/Link'
 import { IconDiscord } from 'components/OSIcons/Icons'
-
-// ─────────────────────────────────────────────
-// Download CTA Button
-// ─────────────────────────────────────────────
-
-function DownloadButton({ autoFocus = false }: { autoFocus?: boolean }) {
-    const posthog = usePostHog()
-    const selectedProduct = useProduct({ handle: 'posthog_code' })
-    const { setConfetti } = useApp()
-    const [email, setEmail] = useState('')
-    const [submitted, setSubmitted] = useState(false)
-    const inputRef = useRef<HTMLInputElement>(null)
-
-    const handleSubmit = (e: React.FormEvent) => {
-        e.preventDefault()
-        if (!email) return
-        posthog?.capture('subscribe_to_product_updates', { email, selectedProduct })
-        setConfetti(true)
-        setSubmitted(true)
-    }
-
-    if (submitted) {
-        return (
-            <p className="text-sm mt-0 mb-4 border border-green rounded-md p-3 bg-green/10">
-                <strong>You&apos;re on the list!</strong>
-                <br />
-                We&apos;ll let you know when <span className="inline-block">PostHog Code</span> is ready.
-                <br />
-                <br />
-                <Link
-                    className="group flex items-center gap-1 text-sm font-medium"
-                    to="https://discord.com/invite/E9xV2WnR98"
-                    externalNoIcon
-                >
-                    <IconDiscord className="size-6 text-secondary group-hover:text-primary" />
-                    <span className="group-hover:underline">Join our Discord</span>
-                </Link>
-            </p>
-        )
-    }
-
-    return (
-        <form onSubmit={handleSubmit} className="space-y-2">
-            <h3 className="text-lg font-bold mb-2">Join the waitlist</h3>
-            <Input
-                ref={inputRef}
-                autoFocus={autoFocus}
-                label="Email"
-                type="email"
-                size="md"
-                direction="column"
-                showLabel={false}
-                placeholder="Email address"
-                value={email}
-                onChange={(e: React.ChangeEvent<HTMLInputElement>) => setEmail(e.target.value)}
-                required
-            />
-            <OSButton variant="primary" size="md" width="full" onClick={handleSubmit}>
-                Get updates
-            </OSButton>
-        </form>
-    )
-}
+import { WaitlistForm } from 'components/WaitlistForm'
 
 // ─────────────────────────────────────────────
 // Section label ("The old way", "The PostHog way")
@@ -755,7 +663,7 @@ function HeroSection() {
 
                     <div className="@container max-w-sm">
                         {showForm ? (
-                            <DownloadButton autoFocus />
+                            <WaitlistForm autoFocus />
                         ) : (
                             <>
                                 <OSButton variant="primary" size="lg" onClick={() => setShowForm(true)}>
@@ -1320,7 +1228,7 @@ const TLDR = () => {
             <h2 className="text-2xl font-bold mb-2">Try it</h2>
             <p>PostHog Code is launching in Spring 2026. Join the waitlist to be the first to try it.</p>
             <div className="max-w-lg @container bg-blue/10 border border-blue rounded-md px-8 py-6 shadow-xl">
-                <DownloadButton />
+                <WaitlistForm />
             </div>
         </section>
     )
